Runtime statistics were never cleared, making them useless during unit-tests
