+++ util.py
@@ -15,6 +15,8 @@ from twisted.internet.task import LoopingCall
 from twisted.python import failure
 from twisted.python.threadable import isInIOThread
 
+from .statistics import RuntimeStatistic
+
 
 logger = logging.getLogger(__name__)
 
@@ -108,37 +110,6 @@ else:
     def attach_profiler(func):
         return func
 
-
-class RuntimeStatistic(object):
-
-    def __init__(self):
-        self._count = 0
-        self._duration = 0.0
-
-    @property
-    def count(self):
-        " Returns the number of times a method was called. "
-        return self._count
-
-    @property
-    def duration(self):
-        " Returns the cumulative time spent in a method. "
-        return self._duration
-
-    @property
-    def average(self):
-        " Returns the average time spent in a method. "
-        return self._duration / self._count
-
-    def increment(self, duration):
-        " Increase self.count with 1 and self.duration with DURATION. "
-        assert isinstance(duration, float), type(duration)
-        self._duration += duration
-        self._count += 1
-
-    def get_dict(self, **kargs):
-        " Returns a dictionary with the statistics. "
-        return dict(count=self.count, duration=self.duration, average=self.average, **kargs)
 
 _runtime_statistics = defaultdict(RuntimeStatistic)
